pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "capstone-a1"
version = "0.1.0"
description = "Factuality assessment for news articles using Google ADK and Gemini"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
authors = [{ name = "Capstone Team" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "scikit-learn>=1.2.0",
    "joblib>=1.2.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "google-adk>=0.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Repository = "https://github.com/yourusername/Capstone-A1"
Documentation = "https://github.com/yourusername/Capstone-A1#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]
exclude = ["src.scripts*", "src.tests*"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(\.git|__pycache__|venv|\.venv|build|dist)/
'''

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "I", "W"]
ignore = ["E501"]
exclude = [".git", "__pycache__", "venv", ".venv", "build", "dist"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
